Extract a custom hook
